move MNIST to PL
