+++ tests/base/__init__.py
@@ -1,4 +1,3 @@
 """Models for testing."""
 
-from tests.base.datasets import TrialMNIST
 from tests.base.model_template import EvalModelTemplate
